+++ __tests__/view/view.oldUser.js
@@ -60,5 +60,11 @@
 				'app/views/viewName/languages/en.json'
 			]);
 		});
+
+		it('create main less file', () => {
+			assert.file([
+				'app/styles/less/' + appName + '.less'
+			]);
+		});
 	});
 })();