+++ generators/view/writing/stylesFiles.js
@@ -12,6 +12,8 @@
 
 	const utils = require('../../common/utils.js');
 
+	const path = require('path');
+
 	module.exports = $that => {
 		const viewsPath  = utils.getViewsPath($that);
 		const filePrefix = viewsPath + $that.viewNameCamel + '/styles/' + $that.viewNameCamel + '.';
@@ -27,6 +29,24 @@
 		utils.copyFileTpl($that, 'styles/main.less', config, filePrefix + 'less');
 		utils.copyFileTpl($that, 'styles/mixins.less', config, filePrefix + 'mixins.less');
 		utils.copyFileTpl($that, 'styles/variables.less', config, filePrefix + 'variables.less');
+
+		const styleFilePath = path.join(process.cwd(), 'app/styles/less', $that.config.get('appNameKebab') + '.less');
+		/* istanbul ignore if */
+		if ($that.fs.exists(styleFilePath)) {
+			write();
+		}
+		else {
+			$that.fs.write(styleFilePath, '');
+			write();
+		}
+
+		function write() {
+			let styleFile = $that.fs.read(styleFilePath);
+			styleFile += '\n';
+			styleFile += '@import (less) "../../views/' + $that.viewPath || '';
+			styleFile += $that.viewNameCamel + '/styles/' + $that.viewNameCamel + '.less';
+			$that.fs.write(styleFilePath, styleFile);
+		}
 	};
 
 })();